Revert "Feat quantity"
