+++ app/reducers/index.jsx
@@ -3,7 +3,6 @@ import { combineReducers } from 'redux'
 const rootReducer = combineReducers({
   auth: require('./auth').default,
   products: require('../components/AllProducts/reducers/products').default,
-  manageCart: require('../components/ManageCart/reducers/manageCart').default,
 })
 
 export default rootReducer
