+++ db/models/index.js
@@ -20,8 +20,6 @@ const app = require('APP')
       Products: require('./products'),
       Types: require('./productTypes'),
       Holiday: require('./holidays'),
-      Cart: require('./cart'),
-      InCart: require('./inCart'),
       // ---------- Add new models here ----------
       Order: require('./order')
     }
