+++ server/api.js
@@ -6,6 +6,7 @@ api
   .get('/heartbeat', (req, res) => res.send({ok: true}))
   .use('/auth', require('./auth'))
   .use('/users', require('./users'))
+  .use('/cart', require('./cart'))
   .use('/products', require('./products'))
   .use('/checkout', require('./checkout'))
