+++ server/users.js
@@ -2,7 +2,6 @@
 
 const db = require('APP/db')
 const User = db.model('users')
-const Cart = db.model('cart')
 
 const {mustBeLoggedIn, forbidden} = require('./auth.filters')
 
@@ -29,22 +28,4 @@ module.exports = require('express').Router()
     (req, res, next) =>
       User.findById(req.params.id)
       .then(user => res.json(user))
-      .catch(next))
-  .get('/:id/cart',
-    mustBeLoggedIn,
-    (req, res, next) =>
-      User.findById(req.params.id)
-      .then(user => {
-        Cart.findOrCreate({where: {userId: req.params.id}})
-      })
-      .then(cart => res.status(201).json(cart))
-      .catch(next))
-  .get('/:id/cart/products',
-    mustBeLoggedIn,
-    (req, res, next) =>
-      User.findById(req.params.id)
-      .then(user => {
-        Cart.findOrCreate({where: {userId: req.params.id}})
-      })
-      .then(cart => res.json(cart.products))
       .catch(next))
